refactor: Extract ImageDiscoveryWorker to separate module
- Move ImageDiscoveryWorker from main_window.py to workers/image_discovery_worker.py
- Reduce main_window.py from 9,252 to 9,191 lines
- First step in incremental refactoring of monolithic main window
