+++ src/lazylabel/ui/workers/__init__.py
@@ -1,0 +1,5 @@
+"""Worker thread classes for background operations."""
+
+from .image_discovery_worker import ImageDiscoveryWorker
+
+__all__ = ["ImageDiscoveryWorker"]
